add go-astilectron code
