+++ .gitignore
@@ -1,0 +1,4 @@
+# production
+/output
+/.idea
+.idea
