+++ bundler.json
@@ -1,0 +1,9 @@
+{
+  "app_name": "Slave's Work",
+
+  "environments": [
+    {"arch": "amd64", "os": "linux"},
+    {"arch": "amd64", "os": "windows"},
+    {"arch": "amd64", "os": "darwin"}
+  ]
+}
