+++ main.go
@@ -1,171 +1,114 @@
 package main
 
 import (
-	"bufio"
-	"context"
-	"crypto/rand"
-	"errors"
-	"flag"
 	"fmt"
-	"io"
+	"github.com/asticode/go-astikit"
+	"github.com/asticode/go-astilectron"
 	"log"
-	"time"
-
-	"github.com/libp2p/go-libp2p"
-	"github.com/libp2p/go-libp2p-core/crypto"
-	"github.com/libp2p/go-libp2p-core/network"
-	"github.com/libp2p/go-libp2p-core/peer"
-	peerstore "github.com/libp2p/go-libp2p-peerstore"
-	"github.com/multiformats/go-multiaddr"
-	"github.com/shirou/gopsutil/cpu"
 )
 
-var errUsageOfCPU error = errors.New("Unable to measure CPU usage.")
+const StaticPath string = "static"
 
 func main() {
-	sourcePort := flag.Int("sp", 4444, "Source port number")
-	dest := flag.String("d", "", "Destination multiaddr string")
+	logger := setLogger()
+	app := createAstilectronAPP(logger)
 
-	flag.Parse()
+	err := app.Start()
+	errorChecker(err, logger, "main: starting astilectron failed: %w")
+	defer app.Close()
 
-	// Creates a new RSA key pair for this host.
-	var r io.Reader = rand.Reader
-	prvKey, _, err := crypto.GenerateKeyPairWithReader(crypto.RSA, 2048, r)
+	window, err := app.NewWindow(StaticPath+"/index.html", &astilectron.WindowOptions{
+		Center: astikit.BoolPtr(true),
+		Height: astikit.IntPtr(700),
+		Width:  astikit.IntPtr(700),
+	})
+	errorChecker(err, logger, "main: new window failed: %w")
+
+	err = window.Create()
+	errorChecker(err, logger, "main: creating window failed: %w")
+
+	newMenu(app, window)
+
+	// Blocking pattern
+	app.Wait()
+}
+
+// Set logger
+func setLogger() *log.Logger {
+	return log.New(log.Writer(), log.Prefix(), log.Flags())
+}
+
+// Create astilectron
+func createAstilectronAPP(logger *log.Logger) *astilectron.Astilectron {
+	a, err := astilectron.New(logger, astilectron.Options{
+		AppName:           "Test",
+		BaseDirectoryPath: "",
+	})
+	errorChecker(err, logger, "main: creating astilectron failed: %w")
+	return a
+}
+
+// Error Checker
+func errorChecker(err error, logger *log.Logger, message string){
 	if err != nil {
-		panic(err)
+		logger.Fatal(fmt.Errorf(message, err))
 	}
+}
 
-	// 0.0.0.0 will listen on any interface device.
-	sourceMultiAddr, _ := multiaddr.NewMultiaddr(fmt.Sprintf("/ip4/0.0.0.0/tcp/%d", *sourcePort))
+// NewMenu
+func newMenu(a *astilectron.Astilectron, w *astilectron.Window) {
+	menu := a.NewMenu([]*astilectron.MenuItemOptions{
+		{
+			Label: astikit.StrPtr("Separator"),
+			SubMenu: []*astilectron.MenuItemOptions{
+				{Label: astikit.StrPtr("Normal 1")},
+				{Label: astikit.StrPtr("Normal 2"),
+					OnClick: func(e astilectron.Event) (deleteListener bool) {
+						w.SendMessage("hello", func(m *astilectron.EventMessage) {
+							// Unmarshal
+							var s string
+							m.Unmarshal(&s)
 
-	// libp2p.New constructs a new libp2p Host.
-	// Other options can be added here.
-	host, err := libp2p.New(
-		context.Background(),
-		libp2p.ListenAddrs(sourceMultiAddr),
-		libp2p.Identity(prvKey),	
-	)
+							// Process message
+							log.Printf("received %s\n", s)
+						})
+						return
+					},
+				},
+				{Type: astilectron.MenuItemTypeSeparator},
+				{Label: astikit.StrPtr("dev tool"), OnClick: func(e astilectron.Event) (deleteListener bool) {
+					w.OpenDevTools()
+					return
+				}},
+			},
+		},
+		{
+			Label: astikit.StrPtr("Checkbox"),
+			SubMenu: []*astilectron.MenuItemOptions{
+				{Checked: astikit.BoolPtr(true), Label: astikit.StrPtr("Checkbox 1"), Type: astilectron.MenuItemTypeCheckbox},
+				{Label: astikit.StrPtr("Checkbox 2"), Type: astilectron.MenuItemTypeCheckbox},
+				{Label: astikit.StrPtr("Checkbox 3"), Type: astilectron.MenuItemTypeCheckbox},
+			},
+		},
+		{
+			Label: astikit.StrPtr("Radio"),
+			SubMenu: []*astilectron.MenuItemOptions{
+				{Checked: astikit.BoolPtr(true), Label: astikit.StrPtr("Radio 1"), Type: astilectron.MenuItemTypeRadio},
+				{Label: astikit.StrPtr("Radio 2"), Type: astilectron.MenuItemTypeRadio},
+				{Label: astikit.StrPtr("Radio 3"), Type: astilectron.MenuItemTypeRadio},
+			},
+		},
+		{
+			Label: astikit.StrPtr("Roles"),
+			SubMenu: []*astilectron.MenuItemOptions{
+				{Label: astikit.StrPtr("Minimize"), Role: astilectron.MenuItemRoleMinimize},
+				{Label: astikit.StrPtr("Close"), Role: astilectron.MenuItemRoleClose},
+			},
+		},
+	})
 
+	err := menu.Create()
 	if err != nil {
-		panic(err)
+		fmt.Println("Menu Error!")
 	}
-
-	if *dest == "" {
-		// Set a function as stream handler.
-		// This function is called when a peer connects, and starts a stream with this protocol.
-		// Only applies on the receiving side.
-		host.SetStreamHandler("/chat/1.0.0", handleStream)
-
-		// Let's get the actual TCP port from our listen multiaddr, in case we're using 0 (default; random available port).
-		var port string
-		for _, la := range host.Network().ListenAddresses() {
-			if p, err := la.ValueForProtocol(multiaddr.P_TCP); err == nil {
-				port = p
-				break
-			}
-		}
-
-		if port == "" {
-			panic("was not able to find actual local port")
-		}
-
-		fmt.Printf("Run './chat -d /ip4/127.0.0.1/tcp/%v/p2p/%s' on another console.\n", port, host.ID().Pretty())
-		fmt.Println("You can replace 127.0.0.1 with public IP as well.")
-		fmt.Printf("\nWaiting for incoming connection\n\n")
-
-		// Hang forever
-		<-make(chan struct{})
-	} else {
-		fmt.Println("This node's multiaddresses:")
-		for _, la := range host.Addrs() {
-			fmt.Printf(" - %v\n", la)
-		}
-		fmt.Println()
-
-		// Turn the destination into a multiaddr.
-		maddr, err := multiaddr.NewMultiaddr(*dest)
-		if err != nil {
-			log.Fatalln(err)
-		}
-
-		// Extract the peer ID from the multiaddr.
-		info, err := peer.AddrInfoFromP2pAddr(maddr)
-		if err != nil {
-			log.Fatalln(err)
-		}
-
-		// Add the destination's peer multiaddress in the peerstore.
-		// This will be used during connection and stream creation by libp2p.
-		host.Peerstore().AddAddrs(info.ID, info.Addrs, peerstore.PermanentAddrTTL)
-
-		// Start a stream with the destination.
-		// Multiaddress of the destination peer is fetched from the peerstore using 'peerId'.
-		s, err := host.NewStream(context.Background(), info.ID, "/chat/1.0.0")
-		if err != nil {
-			panic(err)
-		}
-
-		// Create a buffered stream so that read and writes are non blocking.
-		rw := bufio.NewReadWriter(bufio.NewReader(s), bufio.NewWriter(s))
-
-		// Create a thread to read and write data.
-		go writeData(rw)
-
-		// Hang forever.
-		select {}
-	}
-}
-
-func handleStream(s network.Stream) {
-	log.Println("Got a new stream!")
-
-	// Create a buffer stream for non blocking read and write.
-	rw := bufio.NewReadWriter(bufio.NewReader(s), bufio.NewWriter(s))
-
-	go readData(rw)
-
-	// stream 's' will stay open until you close it (or the other side closes it).
-}
-
-func readData(rw *bufio.ReadWriter) {
-	for {
-		str, _ := rw.ReadString('\n')
-
-		if str == "" {
-			return
-		}
-		if str != "\n" {
-			// Green console colour: 	\x1b[32m
-			// Reset console colour: 	\x1b[0m
-			fmt.Println(str)
-		}
-
-	}
-}
-
-func writeData(rw *bufio.ReadWriter) {
-	for {
-		usages, err := cpuUsage()
-
-		if err != nil {
-			panic(err)
-		}
-
-		usageStr := ""
-		for i := 0; i < len(usages); i++ {
-			usageStr += fmt.Sprint(i) + " : " + fmt.Sprint(usages[i]) + "%\n"
-		}
-
-		rw.WriteString(usageStr)
-		rw.Flush()
-	}
-
-}
-
-func cpuUsage() ([]float64, error) {
-	percent, err := cpu.Percent(time.Duration(1)*time.Second, false)
-	if err != nil {
-		return nil, errUsageOfCPU
-	}
-	return percent, nil
 }
