+++ sendCPUinfo/main.go
@@ -1,0 +1,171 @@
+package main
+
+import (
+	"bufio"
+	"context"
+	"crypto/rand"
+	"errors"
+	"flag"
+	"fmt"
+	"io"
+	"log"
+	"time"
+
+	"github.com/libp2p/go-libp2p"
+	"github.com/libp2p/go-libp2p-core/crypto"
+	"github.com/libp2p/go-libp2p-core/network"
+	"github.com/libp2p/go-libp2p-core/peer"
+	peerstore "github.com/libp2p/go-libp2p-peerstore"
+	"github.com/multiformats/go-multiaddr"
+	"github.com/shirou/gopsutil/cpu"
+)
+
+var errUsageOfCPU error = errors.New("Unable to measure CPU usage.")
+
+func main() {
+	sourcePort := flag.Int("sp", 4444, "Source port number")
+	dest := flag.String("d", "", "Destination multiaddr string")
+
+	flag.Parse()
+
+	// Creates a new RSA key pair for this host.
+	var r io.Reader = rand.Reader
+	prvKey, _, err := crypto.GenerateKeyPairWithReader(crypto.RSA, 2048, r)
+	if err != nil {
+		panic(err)
+	}
+
+	// 0.0.0.0 will listen on any interface device.
+	sourceMultiAddr, _ := multiaddr.NewMultiaddr(fmt.Sprintf("/ip4/0.0.0.0/tcp/%d", *sourcePort))
+
+	// libp2p.New constructs a new libp2p Host.
+	// Other options can be added here.
+	host, err := libp2p.New(
+		context.Background(),
+		libp2p.ListenAddrs(sourceMultiAddr),
+		libp2p.Identity(prvKey),	
+	)
+
+	if err != nil {
+		panic(err)
+	}
+
+	if *dest == "" {
+		// Set a function as stream handler.
+		// This function is called when a peer connects, and starts a stream with this protocol.
+		// Only applies on the receiving side.
+		host.SetStreamHandler("/chat/1.0.0", handleStream)
+
+		// Let's get the actual TCP port from our listen multiaddr, in case we're using 0 (default; random available port).
+		var port string
+		for _, la := range host.Network().ListenAddresses() {
+			if p, err := la.ValueForProtocol(multiaddr.P_TCP); err == nil {
+				port = p
+				break
+			}
+		}
+
+		if port == "" {
+			panic("was not able to find actual local port")
+		}
+
+		fmt.Printf("Run './chat -d /ip4/127.0.0.1/tcp/%v/p2p/%s' on another console.\n", port, host.ID().Pretty())
+		fmt.Println("You can replace 127.0.0.1 with public IP as well.")
+		fmt.Printf("\nWaiting for incoming connection\n\n")
+
+		// Hang forever
+		<-make(chan struct{})
+	} else {
+		fmt.Println("This node's multiaddresses:")
+		for _, la := range host.Addrs() {
+			fmt.Printf(" - %v\n", la)
+		}
+		fmt.Println()
+
+		// Turn the destination into a multiaddr.
+		maddr, err := multiaddr.NewMultiaddr(*dest)
+		if err != nil {
+			log.Fatalln(err)
+		}
+
+		// Extract the peer ID from the multiaddr.
+		info, err := peer.AddrInfoFromP2pAddr(maddr)
+		if err != nil {
+			log.Fatalln(err)
+		}
+
+		// Add the destination's peer multiaddress in the peerstore.
+		// This will be used during connection and stream creation by libp2p.
+		host.Peerstore().AddAddrs(info.ID, info.Addrs, peerstore.PermanentAddrTTL)
+
+		// Start a stream with the destination.
+		// Multiaddress of the destination peer is fetched from the peerstore using 'peerId'.
+		s, err := host.NewStream(context.Background(), info.ID, "/chat/1.0.0")
+		if err != nil {
+			panic(err)
+		}
+
+		// Create a buffered stream so that read and writes are non blocking.
+		rw := bufio.NewReadWriter(bufio.NewReader(s), bufio.NewWriter(s))
+
+		// Create a thread to read and write data.
+		go writeData(rw)
+
+		// Hang forever.
+		select {}
+	}
+}
+
+func handleStream(s network.Stream) {
+	log.Println("Got a new stream!")
+
+	// Create a buffer stream for non blocking read and write.
+	rw := bufio.NewReadWriter(bufio.NewReader(s), bufio.NewWriter(s))
+
+	go readData(rw)
+
+	// stream 's' will stay open until you close it (or the other side closes it).
+}
+
+func readData(rw *bufio.ReadWriter) {
+	for {
+		str, _ := rw.ReadString('\n')
+
+		if str == "" {
+			return
+		}
+		if str != "\n" {
+			// Green console colour: 	\x1b[32m
+			// Reset console colour: 	\x1b[0m
+			fmt.Println(str)
+		}
+
+	}
+}
+
+func writeData(rw *bufio.ReadWriter) {
+	for {
+		usages, err := cpuUsage()
+
+		if err != nil {
+			panic(err)
+		}
+
+		usageStr := ""
+		for i := 0; i < len(usages); i++ {
+			usageStr += fmt.Sprint(i) + " : " + fmt.Sprint(usages[i]) + "%\n"
+		}
+
+		rw.WriteString(usageStr)
+		rw.Flush()
+	}
+
+}
+
+func cpuUsage() ([]float64, error) {
+	percent, err := cpu.Percent(time.Duration(1)*time.Second, false)
+	if err != nil {
+		return nil, errUsageOfCPU
+	}
+	return percent, nil
+}
