+++ static/script/event.js
@@ -1,0 +1,6 @@
+
+function ClientButtonEvent(){
+    astilectron.sendMessage("hello", function(message) {
+        console.log("received " + message)
+    });
+}
